Consolidate build infra: code-safety tooling and correctness fixes
Add a comprehensive code-safety, security, and robustness toolchain and fix the
correctness bugs it surfaced.

Correctness: RT-safe atomic wait/notify for Semaphore, Latch, and AsyncLogBackend
lost-wakeups; undefined-behaviour fixes in the MNA solver and scheduler loader;
data races in concurrency and protocol tests; executive shutdown-mode inference
from timing args; and bounded CLI parses that clamp instead of silently
truncating into 32-bit fields.

Tooling: composable sanitizers (asan/tsan/ubsan) plus RealtimeSanitizer; gated
clang-tidy and cppcheck; trivy, gitleaks, osv-scanner, semgrep, CodeQL, and
Dependabot; a hardened release build; MC/DC coverage with faster reports; a check
registry driving a matrix nightly with tiered CI; and a curated -Werror gate.

Verified: full build clean under -Werror, all unit tests pass (C++/Rust/Python),
asan-ubsan/static/docs gates clean, and the Monte Carlo, Circuit, Edge (GPU), and
Time demos reproduce their archived results.

+++ apps/apex_hil_demo/driver/inc/HilDriver.hpp
@@ -422,8 +422,8 @@ private:
     // Verify CRC-16/XMODEM
     const std::size_t PAYLOAD_LEN = len - 2;
     apex::checksums::crc::Crc16XmodemTable crc;
-    const std::uint16_t EXPECTED =
-        (static_cast<std::uint16_t>(data[PAYLOAD_LEN]) << 8) | data[PAYLOAD_LEN + 1];
+    const std::uint16_t EXPECTED = static_cast<std::uint16_t>(
+        (static_cast<std::uint16_t>(data[PAYLOAD_LEN]) << 8) | data[PAYLOAD_LEN + 1]);
     std::uint16_t computed = 0;
     crc.calculate(data, PAYLOAD_LEN, computed);
